+++ custom_components/truenas/const.py
@@ -4,7 +4,13 @@ import voluptuous as vol
 from homeassistant.const import Platform
 from homeassistant.helpers import config_validation as cv
 
-PLATFORMS = [Platform.SENSOR, Platform.BINARY_SENSOR, Platform.UPDATE, Platform.SWITCH]
+PLATFORMS = [
+    Platform.SENSOR,
+    Platform.BINARY_SENSOR,
+    Platform.BUTTON,
+    Platform.UPDATE,
+    Platform.SWITCH,
+]
 
 DOMAIN = "truenas"
 DEFAULT_NAME = "root"
@@ -93,6 +99,9 @@ SCHEMA_SERVICE_REPLICATION_RUN = {}
 
 SERVICE_DATASET_SNAPSHOT = "dataset_snapshot"
 SCHEMA_SERVICE_DATASET_SNAPSHOT = {}
+
+SERVICE_SNAPSHOTTASK_RUN = "snapshottask_run"
+SCHEMA_SERVICE_SNAPSHOTTASK_RUN = {}
 
 SERVICE_SYSTEM_REBOOT = "system_reboot"
 SCHEMA_SERVICE_SYSTEM_REBOOT = {}